% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{build_B_spline_basis}
\alias{build_B_spline_basis}
\title{Construct B-spline basis}
\usage{
build_B_spline_basis(X, Y, kr, deg, diff, type)
}
\arguments{
\item{X}{vector with ages}

\item{Y}{vector with years}

\item{kr}{Knot ratio. Number of internal intervals used for defining 1 knot in 
B-spline basis construction. See \code{\link{MortSmooth_bbase}}.
Default \code{2} in \code{control.pclm} and \code{7} in
\code{control.pclm2D}. In the two-dimensional model \code{kr} applies to
both axes, so it must not exceed the length of the shorter one: a panel of
fewer than \code{kr} years has no internal knot and is rejected.}

\item{deg}{Degree of the splines needed to create equally-spaced B-splines 
basis over an abscissa of data. Default: 3, a cubic spline. Must be an
integer of at least 2.}

\item{diff}{An integer indicating the order of differences of the components 
of PCLM coefficients. Default value: 2.}

\item{type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}
}
\description{
This is an internal function which constructs B-spline basis to be used in
pclm estimation
}
\seealso{
\code{\link{MortSmooth_bbase}}
}
\keyword{internal}
