% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{create.artificial.bin}
\alias{create.artificial.bin}
\title{Create an additional bin with a small value at the end. 
Improves convergence.}
\usage{
create.artificial.bin(i, vy = 1, vo = 1.01)
}
\arguments{
\item{i}{A list of input values corresponding to pclm or pclm2D}

\item{vy}{Numerical value of the bin created for 'y' input}

\item{vo}{Numerical values of the bin created for 'offset' input (if the case)}
}
\description{
Create an additional bin with a small value at the end. 
Improves convergence.
}
\keyword{internal}
