% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{delete.artificial.bin}
\alias{delete.artificial.bin}
\title{Delete from results the last group added artificially in pclm and pclm2D}
\usage{
delete.artificial.bin(M)
}
\arguments{
\item{M}{A pclm.fit object}
}
\description{
Delete from results the last group added artificially in pclm and pclm2D
}
\keyword{internal}
