% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frac}
\alias{frac}
\title{Extract Fractional Part of a Number}
\usage{
frac(x)
}
\arguments{
\item{x}{A numeric value, vector or matrix}
}
\description{
Extract Fractional Part of a Number
}
\keyword{internal}
