% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{map.bins}
\alias{map.bins}
\title{Map groups and borders}
\usage{
map.bins(x, nlast, out.step)
}
\arguments{
\item{x}{Vector containing the starting values of the input intervals/bins.
For example: if we have 3 bins \code{[0,5), [5,10) and [10, 15)},
\code{x} will be defined by the vector: \code{c(0, 5, 10)}.}

\item{nlast}{Length of the last interval. In the example above \code{nlast} would be 5.}

\item{out.step}{Length of estimated intervals in output. 
Values between 0.1 and 1 are accepted. Default: 1.}
}
\description{
We assume no missing values between the bins
}
\keyword{internal}
