% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalitySmooth_fun.R
\name{MortSmooth_tpower}
\alias{MortSmooth_tpower}
\title{Truncated p-th Power Function}
\usage{
MortSmooth_tpower(x, tt, p)
}
\arguments{
\item{x}{for the abscissa of data;}

\item{tt}{vector of truncation points;}

\item{p}{degree of the power}
}
\description{
This is an internal function of package MortalitySmooth which constructs a 
truncated p-th power function along an abscissa within the function 
MortSmooth_bbase
}
\details{
Internal function used in MortSmooth_bbase. 
The vector t contains the knots. The simplest system of truncated power 
functions uses p = 0 and it consists of step functions with jumps of size 1 
at the truncation points t.
}
\author{
Carlo G Camarda
}
\keyword{internal}
