% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_optim.R
\name{objective_fn}
\alias{objective_fn}
\title{Objective function}
\usage{
objective_fn(L, opt.method, tol = 1e-05, ...)
}
\arguments{
\item{L}{lambda}

\item{opt.method}{Selection criterion of the model.
Possible values are \code{"AIC"} and \code{"BIC"}. Default: \code{"BIC"}.}

\item{tol}{Relative tolerance in PCLM fitting procedure. Default: 0.1\% i.e. 
the estimated aggregate bins should be in the 0.1\% error margin.}

\item{...}{all the other pclm.fit arguments}
}
\description{
Objective function
}
\keyword{internal}
