% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_optim.R
\name{ofun}
\alias{ofun}
\title{Objective function}
\usage{
ofun(L, I, type)
}
\arguments{
\item{L}{lambda.hat}

\item{I}{Input object from pclm function}

\item{type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}
}
\description{
Objective function
}
\keyword{internal}
