% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_optim.R
\name{optimize_par1D}
\alias{optimize_par1D}
\title{Optimize Smoothing Parameters - 2nd method}
\usage{
optimize_par1D(x, y, nlast, offset, show, out.step, control, pclm.type)
}
\arguments{
\item{x}{Vector containing the starting values of the input intervals/bins.
For example: if we have 3 bins \code{[0,5), [5,10) and [10, 15)},
\code{x} will be defined by the vector: \code{c(0, 5, 10)}.}

\item{y}{Vector with counts to be ungrouped. It must have the same dimension as \code{x}.}

\item{nlast}{Length of the last interval. In the example above \code{nlast} would be 5.}

\item{offset}{Optional offset term to calculate smooth mortality rates. 
A vector of the same length as x and y. See \emph{Rizzi et al.(2015)} for 
further details.}

\item{show}{Logical value. Indicates whether a progress bar should be shown or not.
Default: \code{TRUE}.}

\item{out.step}{Length of estimated intervals in output. 
Values between 0.1 and 1 are accepted. Default: 1.}

\item{control}{List with additional parameters. See \code{\link{pclm.control}}.}
}
\description{
Optimize Smoothing Parameters - 2nd method
}
\keyword{internal}
