% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_optim.R
\name{optimize_par2D}
\alias{optimize_par2D}
\title{Optimize Smoothing Parameters
This function optimize searches of \code{lambda, kr} and \code{deg}. 
See \code{\link{pclm.control}} to see what is their meaning. 
The optimization process works in steps. Simultaneous optimization was 
tested and found inefficient. Methods tested: "Nelder-Mead" (optim) 
and "PORT routines" (nlminb).}
\usage{
optimize_par2D(x, y, nlast, offset, show, out.step, control)
}
\arguments{
\item{x}{Vector containing the starting values of the input intervals/bins.
For example: if we have 3 bins \code{[0,5), [5,10) and [10, 15)},
\code{x} will be defined by the vector: \code{c(0, 5, 10)}.}

\item{y}{Vector with counts to be ungrouped. It must have the same dimension as \code{x}.}

\item{nlast}{Length of the last interval. In the example above \code{nlast} would be 5.}

\item{offset}{Optional offset term to calculate smooth mortality rates. 
A vector of the same length as x and y. See \emph{Rizzi et al.(2015)} for 
further details.}

\item{show}{Logical value. Indicates whether a progress bar should be shown or not.
Default: \code{TRUE}.}

\item{out.step}{Length of estimated intervals in output. 
Values between 0.1 and 1 are accepted. Default: 1.}

\item{control}{List with additional parameters: \itemize{
\item{\code{lambda}} -- Smoothing parameter to be used in pclm estimation.
If \code{lambda = NA} an algorithm will find the optimal values.
\item{kr} -- Knot ratio. Number of internal intervals used for defining 
1 knot in B-spline basis construction. See \code{\link{MortSmooth_bbase}}.
\item{\code{deg}} -- Degree of the splines needed to create equally-spaced 
B-splines basis over an abscissa of data.
\item{\code{int.lambda}} -- If \code{lambda} is optimized an interval to be 
searched needs to be specified. Format: vector containing the end-points.
\item{\code{diff}} -- An integer indicating the order of differences of the 
components of PCLM coefficients.
\item{\code{opt.method}} -- Selection criterion of the model.
Possible values are \code{"AIC"} and \code{"BIC"}.
\item{\code{max.iter}} -- Maximal number of iterations used in fitting procedure.
\item{\code{tol}} -- Relative tolerance in PCLM fitting procedure.}}
}
\description{
Optimize Smoothing Parameters
This function optimize searches of \code{lambda, kr} and \code{deg}. 
See \code{\link{pclm.control}} to see what is their meaning. 
The optimization process works in steps. Simultaneous optimization was 
tested and found inefficient. Methods tested: "Nelder-Mead" (optim) 
and "PORT routines" (nlminb).
}
\keyword{internal}
