% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_CI.R
\name{pclm.confidence.dx}
\alias{pclm.confidence.dx}
\title{Compute Confidence Intervals for estimated distribution}
\usage{
pclm.confidence.dx(X)
}
\arguments{
\item{X}{List of inputs from pclm.confidence function}
}
\description{
Compute Confidence Intervals for estimated distribution
}
\keyword{internal}
