% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_CI.R
\name{pclm.confidence.mx}
\alias{pclm.confidence.mx}
\title{Compute Confidence Intervals for estimated hazard}
\usage{
pclm.confidence.mx(X)
}
\arguments{
\item{X}{List of inputs from pclm.confidence function}
}
\description{
Compute Confidence Intervals for estimated hazard
}
\keyword{internal}
