% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_control.R
\name{pclm.control}
\alias{pclm.control}
\title{Auxiliary for Controlling \code{pclm} and \code{pclm2D} Fitting}
\usage{
pclm.control(lambda     = 1,
             kr         = 7,
             deg        = 3,
             int.lambda = c(1, 100),
             int.kr     = c(4, 10),
             int.deg    = c(2, 5),
             diff       = 3,
             opt.method = "BIC",
             max.iter   = 100,
             tol        = 1e-5)
}
\arguments{
\item{lambda}{Smoothing parameter to be used in pclm estimation. Default: 1.
If \code{lambda = NA} an algorithm will find the optimal values.}

\item{kr}{Knot ratio. Number of internal intervals used for defining 1 knot in 
B-spline basis construction. See \code{\link{MortSmooth_bbase}}. Default: 7.
If \code{ndx = NA} an algorithm will find the optimal values.}

\item{deg}{Degree of the splines needed to create equally-spaced B-splines 
basis over an abscissa of data. Default: 3. If \code{deg = NA} an algorithm 
will find the optimal values.}

\item{int.lambda}{If \code{lambda} is optimized an interval to be searched 
needs to be specified. Format: vector containing the end-points.
Default: \code{c(1, 100)}.}

\item{int.kr}{If \code{kr} is optimized an interval to be searched 
needs to be specified. Format: vector containing the end-points.
Default: \code{c(4, 10)}. \code{kr < 5} might slow down considerably
the algorithm in the case of \code{pclm2D} or 'overfit' in the case 
of \code{pclm}. Fell free to play around with it.}

\item{int.deg}{If \code{deg} is optimized an interval to be searched 
needs to be specified. Format: vector containing the end-points.
Default: \code{c(3, 5)}.}

\item{diff}{An integer indicating the order of differences of the components 
of PCLM coefficients. If not sure what it means leave untouched. 
Default value: 3.}

\item{opt.method}{Selection criterion of the model.
Possible values are \code{"AIC"} and \code{"BIC"}. Default: \code{"BIC"}.}

\item{max.iter}{Maximal number of iterations used in fitting procedure.
Default: 100.}

\item{tol}{Tolerance in PCLM fitting procedure.}
}
\value{
List with control parameters.
}
\description{
Auxiliary for Controlling \code{pclm} and \code{pclm2D} Fitting
}
\references{
\enumerate{
\item{Rizzi S, Gampe J, Eilers PHC. \href{https://doi.org/10.1093/aje/kwv020}{
Efficient estimation of smooth distributions from coarsely grouped data.} 
American Journal of Epidemiology, Volume 182, Issue 2, 15 July 2015, Pages 138-147.}
\item{Rizzi S, Thinggaard M, Engholm G, et al. \href{https://doi.org/10.1186/s12874-016-0157-8}{
Comparison of non-parametric methods for ungrouping coarsely aggregated data.} 
BMC Medical Research Methodology. 2016;16:59.}
}
}
\seealso{
\code{\link{pclm}} \code{\link{pclm2D}}
}
