% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_data_imports.R
\name{print.pclm.data}
\alias{print.pclm.data}
\title{Print function for \code{pclm.data}}
\usage{
\method{print}{pclm.data}(x, ...)
}
\arguments{
\item{x}{An \code{pclm.data} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for \code{pclm.data}
}
\keyword{internal}
