% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_1D.R
\name{print.pclm}
\alias{print.pclm}
\title{Print for pclm method}
\usage{
\method{print}{pclm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"pclm"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print for pclm method
}
\keyword{internal}
