% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_2D.R
\name{print.pclm2D}
\alias{print.pclm2D}
\title{Print method for pclm2D}
\usage{
\method{print}{pclm2D}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"pclm2D"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for pclm2D
}
\keyword{internal}
