% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_1D.R
\name{print.summary.pclm}
\alias{print.summary.pclm}
\title{Print for summary.pclm method}
\usage{
\method{print}{summary.pclm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.pclm"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print for summary.pclm method
}
\keyword{internal}
