% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_2D.R
\name{print.summary.pclm2D}
\alias{print.summary.pclm2D}
\title{Print method for summary.pclm2D}
\usage{
\method{print}{summary.pclm2D}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.pclm2D"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for summary.pclm2D
}
\keyword{internal}
