% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup-data.R
\name{print.ungroup.data}
\alias{print.ungroup.data}
\title{Print function for \code{ungroup.data}}
\usage{
\method{print}{ungroup.data}(x, ...)
}
\arguments{
\item{x}{An \code{ungroup.data} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for \code{ungroup.data}
}
\keyword{internal}
