% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_1D.R
\name{residuals.pclm}
\alias{residuals.pclm}
\title{Extract PCLM Deviance Residuals}
\usage{
\method{residuals}{pclm}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model residuals is
    meaningful.}

\item{...}{other arguments.}
}
\value{
Residuals extracted from the object \code{object}.
}
\description{
Extract PCLM Deviance Residuals
}
\examples{
x <- c(0, 1, seq(5, 85, by = 5))
y <- c(294, 66, 32, 44, 170, 284, 287, 293, 361, 600, 998, 
       1572, 2529, 4637, 6161, 7369, 10481, 15293, 39016)
M1 <- pclm(x, y, nlast = 26)

residuals(M1)
}
