% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_nlast}
\alias{resolve_nlast}
\title{Resolve the width of the last interval}
\usage{
resolve_nlast(x, nlast, omega)
}
\arguments{
\item{x}{Vector containing the starting values of the input intervals/bins.}

\item{nlast}{The width of the last interval. See \code{\link{pclm}}.}

\item{omega}{The closing age of the distribution, an alternative to
\code{nlast}.}
}
\description{
A caller may state either the width of the last interval or the closing age
of the distribution. This turns the second into the first.
}
\keyword{internal}
