% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seqlast}
\alias{seqlast}
\title{Sequence function with last value}
\usage{
seqlast(from, to, by)
}
\arguments{
\item{from, to}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{by}{number: increment of the sequence.}
}
\description{
Sequence function with last value
}
\keyword{internal}
