% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{suggest.valid.out.step}
\alias{suggest.valid.out.step}
\title{Suggest values of 'out.step' that do not require an adjustment of 'nlast'}
\usage{
suggest.valid.out.step(len, increment = 0.01)
}
\arguments{
\item{len}{Interval length}

\item{increment}{Increment}
}
\description{
Suggest values of 'out.step' that do not require an adjustment of 'nlast'
}
\keyword{internal}
