% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_1D.R
\name{summary.pclm}
\alias{summary.pclm}
\title{Summary for pclm method}
\usage{
\method{summary}{pclm}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary for pclm method
}
\keyword{internal}
