% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_2D.R
\name{summary.pclm2D}
\alias{summary.pclm2D}
\title{Summary method for pclm2D}
\usage{
\method{summary}{pclm2D}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Generic function used to produce result summaries of the results produced
by \code{\link{pclm2D}}.
}
\keyword{internal}
