% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{build_B_spline_basis}
\alias{build_B_spline_basis}
\title{Construct B-spline basis
This is an internal function which constructs B-spline basis to be used in 
pclm estimation}
\usage{
build_B_spline_basis(X, Y, kr, deg, diff, type)
}
\arguments{
\item{X}{vector with ages}

\item{Y}{vector with years}

\item{type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}
}
\description{
Construct B-spline basis
This is an internal function which constructs B-spline basis to be used in 
pclm estimation
}
\seealso{
\code{\link{MortSmooth_bbase}}
}
\keyword{internal}
