% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{build_B_spline_basis}
\alias{build_B_spline_basis}
\title{Construct B-spline basis
This is an internal function which constructs B-spline basis to be used in 
pclm estimation}
\usage{
build_B_spline_basis(x, kr, deg, diff, lambda, pclm.type)
}
\arguments{
\item{x}{List containing the two numeric vectors for the abscissa of data.}

\item{kr}{Knot ratio. Number of internal intervals used for defining 1 knot in 
B-spline basis construction. See \code{\link{MortSmooth_bbase}}.
If \code{ndx = NA} an algorithm will find the optimal values.}

\item{deg}{Degree of the splines needed to create equally-spaced B-splines 
basis over an abscissa of data. If \code{deg = NA} an algorithm 
will find the optimal values.}

\item{diff}{An integer indicating the order of differences of the components 
of PCLM coefficients. If not sure what it means leave untouched. 
Default value: 3.}

\item{lambda}{Smoothing parameter to be used in pclm estimation.
If \code{lambda = NA} an algorithm will find the optimal values.}

\item{pclm.type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}
}
\description{
Construct B-spline basis
This is an internal function which constructs B-spline basis to be used in 
pclm estimation
}
\seealso{
\code{\link{MortSmooth_bbase}}
}
\keyword{internal}
