% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{build_P_matrix}
\alias{build_P_matrix}
\title{Construct Penalty Matrix}
\usage{
build_P_matrix(BA, BY, lambda, type)
}
\arguments{
\item{BA}{B-spline basis object for age axis}

\item{BY}{B-spline basis object for year axis}

\item{lambda}{Smoothing parameter to be used in pclm estimation.
If \code{lambda = NA} an algorithm will find the optimal values.
\code{pclm} takes a single value. \code{pclm2D} takes two, one for the age
axis and one for the year axis, and either may be \code{NA} to be found by
optimisation while the other is held fixed.}

\item{type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}
}
\description{
Construct Penalty Matrix
}
\keyword{internal}
