% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{compute_standard_errors}
\alias{compute_standard_errors}
\title{Compute Standard Errors and Confidence Intervals}
\usage{
compute_standard_errors(X)
}
\arguments{
\item{X}{An object generated by pclm.fit}
}
\description{
Compute Standard Errors and Confidence Intervals
}
\keyword{internal}
