% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_CI.R
\name{compute_standard_errors}
\alias{compute_standard_errors}
\title{Compute Standard Errors and Confidence Intervals}
\usage{
compute_standard_errors(B, QmQ, QmQP)
}
\arguments{
\item{B}{B-spline basis matrix}

\item{QmQ}{Object generated by pclm.fit}

\item{QmQP}{Object generated by pclm.fit}
}
\description{
Compute Standard Errors and Confidence Intervals
}
\keyword{internal}
