% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_CI.R
\name{pclm.confidence}
\alias{pclm.confidence}
\title{Compute Confidence Intervals for PCLM output}
\usage{
pclm.confidence(fit, out.step, y, SE, ci.level, type, offset)
}
\arguments{
\item{fit}{Fitted values from pclm.fit}

\item{out.step}{Length of estimated intervals in output. 
Values between 0.1 and 1 are accepted. Default: 1.}

\item{y}{Vector with counts to be ungrouped. It must have the same dimension 
as \code{x}.}

\item{SE}{Standard Errors}

\item{ci.level}{Level of significance for computing confidence intervals. 
Default: \code{95}.}

\item{type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}

\item{offset}{Optional offset term to calculate smooth mortality rates. 
A vector of the same length as x and y. See 
\insertCite{rizzi2015;textual}{ungroup} for further details.}
}
\description{
Compute Confidence Intervals for PCLM output
}
\keyword{internal}
