% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{pclm.confidence}
\alias{pclm.confidence}
\title{Compute Standard Errors and Confidence Intervals}
\usage{
pclm.confidence(X, ci.level, pclm.type)
}
\arguments{
\item{X}{An object generated by pclm.fit}

\item{ci.level}{Level of significance for computing confidence intervals. 
Default: \code{0.05}.}
}
\description{
Compute Standard Errors and Confidence Intervals
}
\keyword{internal}
