% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_control.R
\name{pclm.control}
\alias{pclm.control}
\title{Auxiliary for Controlling \code{pclm} Fitting}
\usage{
pclm.control(lambda     = NA,
             kr         = NA,
             deg        = NA,
             int.lambda = c(0, 10^5),
             int.kr     = 4:10,
             int.deg    = 2:5,
             diff       = 2,
             opt.method = "BIC",
             max.iter   = 1e+3,
             tol        = 1e-5)
}
\arguments{
\item{lambda}{Smoothing parameter to be used in pclm estimation.
If \code{lambda = NA} an algorithm will find the optimal values.}

\item{kr}{Knot ratio. Number of internal intervals used for defining 1 knot in 
B-spline basis construction. See \code{\link{MortSmooth_bbase}}.
If \code{ndx = NA} an algorithm will find the optimal values.}

\item{deg}{Degree of the splines needed to create equally-spaced B-splines 
basis over an abscissa of data. If \code{deg = NA} an algorithm 
will find the optimal values.}

\item{int.lambda}{If \code{lambda} is optimized an interval to be searched 
needs to be specified. Format: vector containing the end-points.}

\item{int.kr}{If \code{kr} is optimized, an interval to be searched 
needs to be specified. Format: vector containing the possible values.
Default: \code{4:10}. \code{kr < 4} might slow down considerably
the algorithm in the case of \code{pclm2D} or 'overfit' in the case 
of \code{pclm}. Fell free to play around with it.}

\item{int.deg}{If \code{deg} is optimized, an interval to be searched 
needs to be specified. Format: vector containing the the possible values.
Default: \code{2:5}.}

\item{diff}{An integer indicating the order of differences of the components 
of PCLM coefficients. Default value: 2.}

\item{opt.method}{Selection criterion of the model.
Possible values are \code{"AIC"} and \code{"BIC"}. Default: \code{"BIC"}.}

\item{max.iter}{Maximal number of iterations used in fitting procedure.}

\item{tol}{Tolerance in PCLM fitting procedure.}
}
\value{
List with control parameters.
}
\description{
Auxiliary for Controlling \code{pclm} Fitting
}
\seealso{
\code{\link{pclm}}
}
