% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_control.R
\name{pclm.control}
\alias{pclm.control}
\title{Auxiliary for Controlling \code{pclm} Fitting}
\usage{
pclm.control(lambda     = NA,
             kr         = 6,
             deg        = 3,
             int.lambda = c(0.1, 1e+5),
             diff       = 2,
             opt.method = "BIC",
             max.iter   = 1e+3,
             tol        = 1e-5)
}
\arguments{
\item{lambda}{Smoothing parameter to be used in pclm estimation.
If \code{lambda = NA} an algorithm will find the optimal values.}

\item{kr}{Knot ratio. Number of internal intervals used for defining 1 knot in 
B-spline basis construction. See \code{\link{MortSmooth_bbase}}.}

\item{deg}{Degree of the splines needed to create equally-spaced B-splines 
basis over an abscissa of data.}

\item{int.lambda}{If \code{lambda} is optimized an interval to be searched 
needs to be specified. Format: vector containing the end-points.}

\item{diff}{An integer indicating the order of differences of the components 
of PCLM coefficients. Default value: 2.}

\item{opt.method}{Selection criterion of the model.
Possible values are \code{"AIC"} and \code{"BIC"}. Default: \code{"BIC"}.}

\item{max.iter}{Maximal number of iterations used in fitting procedure.}

\item{tol}{Tolerance in PCLM fitting procedure.}
}
\value{
List with control parameters.
}
\description{
Auxiliary for Controlling \code{pclm} Fitting
}
\seealso{
\code{\link{pclm}}
}
