% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_fit.R
\name{pclm.fit}
\alias{pclm.fit}
\title{Fit PCLM Models}
\usage{
pclm.fit(x, y, nlast, offset, out.step, show, lambda, kr, deg, diff, max.iter,
  tol, type)
}
\arguments{
\item{x}{Vector containing the starting values of the input intervals/bins.
For example: if we have 3 bins \code{[0,5), [5,10) and [10, 15)},
\code{x} will be defined by the vector: \code{c(0, 5, 10)}.}

\item{y}{Vector with counts to be ungrouped. It must have the same dimension as \code{x}.}

\item{nlast}{Length of the last interval. In the example above \code{nlast} would be 5.}

\item{offset}{Optional offset term to calculate smooth mortality rates. 
A vector of the same length as x and y. See \insertCite{rizzi2015;textual}{ungroup} 
for further details.}

\item{out.step}{Length of estimated intervals in output. 
Values between 0.1 and 1 are accepted. Default: 1.}

\item{show}{Logical value. Indicates whether a progress bar should be shown or not.
Default: \code{TRUE}.}

\item{type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}
}
\description{
This is an internal function used to estimate PCLM model. It is used by 
\code{\link{pclm}} and \code{\link{pclm2D}} functions
}
\keyword{internal}
