% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pclm.input.check}
\alias{pclm.input.check}
\title{Validate input values}
\usage{
pclm.input.check(X, pclm.type)
}
\arguments{
\item{X}{A list with input arguments provided in \code{\link{pclm}} function}
}
\description{
Validate input values
}
\keyword{internal}
