% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_utils.R
\name{pclm.input.check}
\alias{pclm.input.check}
\title{Validate input values}
\usage{
pclm.input.check(X, pclm.type)
}
\arguments{
\item{X}{A list with input arguments provided in \code{\link{pclm}} function}

\item{pclm.type}{Type of PCLM model. Options: \code{"1D", "2D"} for 
univariate and two-dimensional model respectively.}
}
\description{
Validate input values
}
\keyword{internal}
