% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_2D.R
\name{pclm2D}
\alias{pclm2D}
\title{Two-Dimensional Penalized Composite Link Model (PCLM-2D)}
\usage{
pclm2D(x, y, nlast, offset = NULL, show = TRUE, ci.level = 0.05,
  out.step = 1, control = list())
}
\arguments{
\item{x}{Vector containing the starting values of the input intervals/bins.
For example: if we have 3 bins \code{[0,5), [5,10) and [10, 15)},
\code{x} will be defined by the vector: \code{c(0, 5, 10)}.}

\item{y}{data.frame with counts to be ungrouped. The number of rows 
should be equal with the length of \code{x}.}

\item{nlast}{Length of the last interval. In the example above \code{nlast = 5}.}

\item{offset}{Optional offset term to calculate smooth mortality rates. 
A vector of the same length as x and y. See \emph{Rizzi et al.(2015)} for 
further details.}

\item{show}{Logical value. Indicates whether a progress bar should be shown or not.
Default: \code{TRUE}.}

\item{ci.level}{Level of significance for computing confidence intervals. 
Default: \code{0.05}.}

\item{out.step}{Length of estimated intervals in output. 
Values between 0.1 and 1 are accepted. Default: 1.}

\item{control}{List with additional parameters. See \code{\link{pclm.control}}.}
}
\value{
The output is a list with the following components:

\item{input}{ A list with arguments provided in input. Saved for convenience.}

\item{fitted}{ The fitted values of the PCLM model.}

\item{ci}{ Confidence intervals around fitted values.}

\item{goodness.of.fit}{ A list containing goodness of fit measures: 
standard errors, AIC and BIC.}

\item{smoothPar}{ Estimated smoothing parameters: \code{lambda, kr} 
and \code{deg}.}

\item{bins.definition}{ Additional values to identify the bins limits and 
location in input and output objects.}

\item{call}{ An unevaluated function call, that is, an unevaluated 
expression which consists of the named function applied to the given arguments.}
}
\description{
Fit two-dimensional penalized composite link model (PCLM-2D), 
e.g. simultaneous ungrouping of age-at-death distributions grouped in age 
classes for adjacent years. The PCLM can be extended to a two-dimensional 
regression problem. This is particularly suitable for mortality analysis 
when mortality surfaces are to be estimated to capture both age-specific 
trajectories of coarsely grouped distributions and time trends.
}
\examples{
# Input data
Dx <- pclm.data$Dx
Ex <- pclm.data$Ex

# Aggregate data
x      <- c(0, 1, seq(5, 85, by = 5))
nlast  <- 26
n      <- c(diff(x), nlast)
Ex$gr  <- Dx$gr <- rep(x, n)
y      <- aggregate(Dx[, 1:35], by = list(Dx$gr), FUN = "sum")[, -1]
offset <- aggregate(Ex[, 1:35], by = list(Ex$gr), FUN = "sum")[, -1]

# Example 1 ---------------------- 
# Fit model and ungroup data using PCLM-2D
P1 <- pclm2D(x, y, nlast)
summary(P1)
# plot(P1)

# Example 2 ---------------------- 
# Ungroup and build a mortality surface
P2 <- pclm2D(x, y, nlast, offset)
summary(P2)

\dontrun{
plot(P2)                      # plot
library(rgl)
snapshot3d("plotP2.jpeg")     # save the plot in jpeg format
aspect3d(x = 1, y = 2, z = 1) # modify the aspect ratio
}

}
\seealso{
\code{\link{plot.pclm2D}}
}
