% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_graphics.R
\name{plot.pclm}
\alias{plot.pclm}
\title{Generic Plot for pclm Class}
\usage{
\method{plot}{pclm}(x, xlab, ylab, ylim, type, lwd, col, legend, legend.position, ...)
}
\arguments{
\item{x}{An object of class \code{\link{pclm}}}

\item{xlab}{a label for the x axis, defaults to a description of \code{x}.}

\item{ylab}{a label for the y axis, defaults to a description of \code{y}.}

\item{ylim}{the y limits of the plot; see \code{xlim} for default and details.}

\item{type}{1-character string giving the type of plot desired. 
The following values are possible, for details, see plot: "p" for points, 
"l" for lines, "b" for both points and lines, "c" for empty points joined 
by lines, "o" for overplotted points and lines, "s" and "S" for stair 
steps and "h" for histogram-like vertical lines. Finally, "n" does not 
produce any points or lines. Default: \code{"l"}.}

\item{lwd}{Line width, a positive number, defaulting to 2.}

\item{col}{Three colors to be used in the plot for observed values, 
fitted values and confidence intervals, in that order. The histogram
default is \code{c("gold2", 2, 4)}; with an offset, which puts the plot on
a log scale, it is \code{c(1, 2, 4)}.}

\item{legend}{Labels for the legend. With three colors the default is
\code{c("Input values", "Fitted values", "Conf. intervals")}.}

\item{legend.position}{Legend position, or the x and y co-ordinates to be 
used to position the legend. The default is chosen from the data: if the
first few bins are taller than the last few the legend goes top right, and
top left otherwise, which keeps it off the peak. The rate plot always
defaults to top left.}

\item{...}{other graphical parameters (see \link{par} for more details).}
}
\description{
Generic Plot for pclm Class
}
\examples{
# See complete examples in pclm help page
}
\seealso{
\code{\link{pclm}}
}
