% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_2D.R
\name{residuals.pclm2D}
\alias{residuals.pclm2D}
\title{Extract PCLM-2D Deviance Residuals}
\usage{
\method{residuals}{pclm2D}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model residuals is
    meaningful.}

\item{...}{other arguments.}
}
\value{
Residuals extracted from the object \code{object}.
}
\description{
Extract PCLM-2D Deviance Residuals
}
