% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suggest.valid.out.step}
\alias{suggest.valid.out.step}
\title{Suggest values of \code{out.step} that do not 
require an adjustment of \code{nlast}}
\usage{
suggest.valid.out.step(len, increment = 0.01)
}
\arguments{
\item{len}{Interval length}

\item{increment}{Increment}
}
\description{
Suggest values of \code{out.step} that do not 
require an adjustment of \code{nlast}
}
\keyword{internal}
