% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup-data.R
\docType{data}
\name{ungroup.data}
\alias{ungroup.data}
\title{Test Dataset in the Package}
\format{
An object of class \code{ungroup.data} of length 2.
}
\source{
\href{https://www.mortality.org}{Human Mortality Database}
}
\usage{
ungroup.data
}
\description{
Dataset containing death counts (Dx) and exposures (Ex)  by age for a 
certain population between 1980 and 2014. The data-set is provided for 
testing purposes only and might be altered and outdated. Download actual 
demographic data free of charge from \insertCite{hmd2018;textual}{ungroup}. 
Once a username and a password is created on the 
\href{https://www.mortality.org}{website} the 
\href{https://CRAN.R-project.org/package=MortalityLaws}{
MortalityLaws} R package can be used to extract data in R format.
\insertNoCite{MortalityLaws160}{ungroup}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[MortalityLaws]{ReadHMD}}
}
\keyword{datasets}
